Ensure current window space is utilised in logging Blocksize and async request optimisations are now the default values.

+++ maverick-synergy-common/src/main/java/com/sshtools/synergy/ssh/ChannelDataWindow.java
@@ -37,7 +37,7 @@ public class ChannelDataWindow {
 		this.windowSpace =  initialWindowSpace;
 	}
 	
-	public synchronized void consume(int count) {
+	public synchronized void consume(long count) {
 		windowSpace = UnsignedInteger32.deduct(windowSpace, count);
 	}
